requirements.txt
tensorflow>=2.0.0
opencv-python>=4.5.0
mediapipe>=0.8.0
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=0.24.0
pyyaml>=5.4.0